refactor(checker): route class static side relation guard
