+++ crates/tsz-checker/src/classes/mod.rs
@@ -7,6 +7,7 @@ pub(crate) mod class_helpers;
 pub(crate) mod class_implements_checker;
 pub(crate) mod class_implements_helpers;
 pub mod class_inheritance;
+pub(crate) mod class_static_side_helpers;
 pub(crate) mod class_summary;
 pub mod constructor_checker;
 pub(crate) mod interface_heritage_display;
